Continuous Markov Chains

There is basic support for continuous Markov chains. All applicable functions also work for continuous chains.

Contents

Continuous Markov Chains

Contents

Index

Documentation

Index

DiscreteMarkovChains.ContinuousMarkovChain

DiscreteMarkovChains.embedded

Documentation

DiscreteMarkovChains.ContinuousMarkovChain — Type

ContinuousMarkovChain(transition_matrix)
ContinuousMarkovChain(state_space, transition_matrix)

Creates a new continuous Markov chain object. This is also known as a Markov jump process.

Note that an irreducible finite continuous-time Markov chain is always positive recurrent and its stationary distribution always exists, is unique and is equal to the limiting distribution.

Arguments

state_space: The names of the states that make up the Markov chain.

transition_matrix: The transition intensity matrix. Also known as the generator matrix.

Examples

The following shows a basic Sunny-Cloudy-Rainy weather model.

using DiscreteMarkovChains
T = [
    -.1 0.1 0;
    0.5 -.8 0.3;
    0.1 0.4 -.5
]
X = ContinuousMarkovChain(["Sunny", "Cloudy", "Rainy"], T)
println(state_space(X))

# output

["Sunny", "Cloudy", "Rainy"]

println(transition_matrix(X))

# output

[-0.1 0.1 0.0; 0.5 -0.8 0.3; 0.1 0.4 -0.5]

References

Wikipedia

DiscreteMarkovChains.embedded — Function

embedded(x)

Arguments

x: some kind of continuous Markov chain.

Returns

The embedded Markov chain for the given continuous Markov chain.

Notes

If the equivalent chain is preffered rather than the embedded chain, then use DiscreteMarkovChain(x).