Communication Functions

Contents

Communication Functions

Contents

Index

Documentation

Index

DiscreteMarkovChains.canonical_form

DiscreteMarkovChains.communication_classes

DiscreteMarkovChains.decompose

DiscreteMarkovChains.periodicities

Documentation

DiscreteMarkovChains.communication_classes — Function

communication_classes(x)

Parameters

x: some kind of Markov chain.

Returns

A tuple containing 2 arrays.

This first array contains C arrays which store the states that communicate.

The second array is an array of Bool where the ith value is true if the ith communication class is recurrent.

DiscreteMarkovChains.periodicities — Function

periodicities(x::AbstractDiscreteMarkovChain)

A more advanced version of communication_classes designed for discrete Markov chains. It is the same as communication_classes but it returns periodicities as well.

Parameters

x: some kind of discrete Markov chain.

Returns

A tuple containing 3 arrays.

This first array contains C arrays which store the states that communicate.

The second array is an array of Bool where the ith value is true if the ith communication class is recurrent.

The third array is the periodicity of each communication class.

DiscreteMarkovChains.decompose — Function

decompose(x)

All transition matrices can be reordered so that we have

\[T = \begin{pmatrix} A & 0\\ B & C \end{pmatrix}\]

Where $A$, $B$ and $C$ are as described below. $0$ is a matrix of zeros.

Parameters

x: some kind of Markov chain.

Returns

A tuple of four values

states: the first value is an array of the new states.

A: the second value is a matrix of recurrent states to recurrent states.

B: the third value is a matrix of transient states to recurrent states.

C: the fourth value is a matrix of transient to transient states.

DiscreteMarkovChains.canonical_form — Function

canonical_form(x)

Reorders the states of the transition matrix of x so that recurrent states appear first and transient states appear last.

Parameters

x: some kind of Markov chain.

Returns

A tuple with the new states and the new transition matrix.