Mean Time Functions

Contents

Mean Time Functions

Contents

Index

Documentation

Index

DiscreteMarkovChains.expected_time_to_absorption

DiscreteMarkovChains.mean_first_passage_time

DiscreteMarkovChains.mean_recurrence_time

Documentation

DiscreteMarkovChains.expected_time_to_absorption — Function

expected_time_to_absorption(x)

The expected number of steps that the process will take while in the transient super class given that the process started in state i.

Parameters

x: some kind of Markov chain.

Returns

A 1D array where element $i$ is the total number of revisits to transient state $i$ before leaving the transient super class.

Note

It is advised to have x in canonical form already. This is to avoid confusion of what states make up each element of the array ouput.

DiscreteMarkovChains.mean_recurrence_time — Function

mean_recurrence_time(x)

This is the expected number of steps for the process to return to state i given that the process started in state i.

Parameters

x: some kind of Markov chain.

Returns

A 1D array where the ith entry is the mean recurrence time of state i.

Note

x must be irreducible (i.e. ergodic).

DiscreteMarkovChains.mean_first_passage_time — Function

mean_first_passage_time(x)

This is the expected number of steps for the process to reach state j given that the process started in state i.

Parameters

x: some kind of Markov chain.

Returns

A matrix where the i,jth entry is the mean recurrence time of state i. Diagonal elements are 0. The diagonals would have represented the mean recurrence time.