Simple Functions

Contents

Simple Functions

Contents

Index

Documentation

Index

DiscreteMarkovChains.is_absorbing

DiscreteMarkovChains.is_ergodic

DiscreteMarkovChains.is_regular

Documentation

DiscreteMarkovChains.is_regular — Function

is_regular(x)

Parameters

x: some kind of Markov chain.

Returns

true if the Markov chain, x, is regular.

DiscreteMarkovChains.is_ergodic — Function

is_ergodic(x)

Parameters

x: some kind of Markov chain.

Returns

true if the Markov chain, x, is ergodic. This is, if every state can be accessed from every other state. Another term for this is irreducible.

DiscreteMarkovChains.is_absorbing — Function

is_absorbing(x)

Parameters

x: some kind of Markov chain.

Returns

true if the Markov chain, x, is an absorbing chain. So the process is guarenteed to be absorbed eventually.