Introduction

Installation

DiscreteMarkovChains should be up on Julia's package registry.

Simply type

Import Pkg
Pkg.add("DiscreteMarkovChains")

DiscreteMarkovChains.DiscreteMarkovChain

DiscreteMarkovChains.canonical_form

DiscreteMarkovChains.communication_classes

DiscreteMarkovChains.decompose

DiscreteMarkovChains.is_absorbing

DiscreteMarkovChains.is_ergodic

DiscreteMarkovChains.is_regular

DiscreteMarkovChains.periodicities

DiscreteMarkovChains.state_space

DiscreteMarkovChains.transition_matrix

Core Types

The main type is DiscreteMarkovChain:

DiscreteMarkovChains.DiscreteMarkovChain — Type

DiscreteMarkovChain(state_space, transition_matrix)

Creates a new discrete Markov chain object.

Parameters

state_space: The names of the states that make up the Markov chain.

transition_matrix: The single step transition probability matrix.

Examples

The following shows a basic Sunny-Cloudy-Rainy weather model.

using DiscreteMarkovChains
T = [
     0.9 0.1 0;
     0.5 0.2 0.3;
     0.1 0.4 0.5
    ]
X = DiscreteMarkovChain(["Sunny", "Cloudy", "Rainy"], T)
print(state_space(X))

# output

["Sunny", "Cloudy", "Rainy"]

DiscreteMarkovChains.state_space — Function

state_space(x)

Parameters

x: some kind of Markov chain.

Returns

The state space of a Markov chain.

DiscreteMarkovChains.transition_matrix — Function

transition_matrix(x)

Returns

The transition matrix of a Markov chain.

Simple Functions

DiscreteMarkovChains.is_regular — Function

is_regular(x)

Returns

true if the Markov chain, x, is regular.

DiscreteMarkovChains.is_ergodic — Function

is_ergodic(x)

Returns

true if the Markov chain, x, is ergodic. This is, if every state can be accessed from every other state. Another term for this is irreducible.

DiscreteMarkovChains.is_absorbing — Function

is_absorbing(x)

Returns

true if the Markov chain, x, is an absorbing chain. So the process is guarenteed to be absorbed eventually.

Communicaton Functions

DiscreteMarkovChains.communication_classes — Function

communication_classes(x)

Parameters

x: some kind of Markov chain.

Returns

A tuple containing 2 arrays.

This first array contains C arrays which store the states that communicate.

The second array is an array of Bool where the ith value is true if the ith communication class is recurrent.

DiscreteMarkovChains.periodicities — Function

periodicities(x::AbstractDiscreteMarkovChain)

A more advanced version of communication_classes designed for discrete Markov chains. It is the same as communication_classes but it returns periodicities as well.

Parameters

x: some kind of discrete Markov chain.

Returns

A tuple containing 3 arrays.

This first array contains C arrays which store the states that communicate.

The second array is an array of Bool where the ith value is true if the ith communication class is recurrent.

The third array is the periodicity of each communication class.

DiscreteMarkovChains.decompose — Function

decompose(x)

Parameters

x: some kind of Markov chain.

Returns

A tuple of four values

The first value is an array of the new states.

The second value is a matrix of recurrent states to recurrent states.

The third value is a matrix of transient states to recurrent states.

The fourth value is a matrix of transient to transient states.

DiscreteMarkovChains.canonical_form — Function

canonical_form(x)

Reorders the states of the transition matrix of x so that we have recurrent states first and transient states last.

Returns

A tuple with the new states and the new transition matrix.