Functions

DiscreteMarkovChains.DiscreteMarkovChain

DiscreteMarkovChains.canonical_form

DiscreteMarkovChains.communication_classes

DiscreteMarkovChains.decompose

DiscreteMarkovChains.exit_probabilities

DiscreteMarkovChains.expected_time_to_absorption

DiscreteMarkovChains.first_passage_probabilities

DiscreteMarkovChains.fundamental_matrix

DiscreteMarkovChains.is_absorbing

DiscreteMarkovChains.is_ergodic

DiscreteMarkovChains.is_regular

DiscreteMarkovChains.periodicities

DiscreteMarkovChains.state_space

DiscreteMarkovChains.stationary_distribution

DiscreteMarkovChains.transition_matrix

Core Types

The main type is DiscreteMarkovChain:

DiscreteMarkovChains.DiscreteMarkovChain — Type

DiscreteMarkovChain(state_space, transition_matrix)

Creates a new discrete Markov chain object.

Parameters

state_space: The names of the states that make up the Markov chain.

transition_matrix: The single step transition probability matrix.

Examples

The following shows a basic Sunny-Cloudy-Rainy weather model.

using DiscreteMarkovChains
T = [
     0.9 0.1 0;
     0.5 0.2 0.3;
     0.1 0.4 0.5
    ]
X = DiscreteMarkovChain(["Sunny", "Cloudy", "Rainy"], T)
print(state_space(X))

# output

["Sunny", "Cloudy", "Rainy"]

DiscreteMarkovChains.state_space — Function

state_space(x)

Parameters

x: some kind of Markov chain.

Returns

The state space of a Markov chain.

DiscreteMarkovChains.transition_matrix — Function

transition_matrix(x)

Parameters

x: some kind of Markov chain.

Returns

The transition matrix of a Markov chain.

Simple Functions

DiscreteMarkovChains.is_regular — Function

is_regular(x)

Parameters

x: some kind of Markov chain.

Returns

true if the Markov chain, x, is regular.

DiscreteMarkovChains.is_ergodic — Function

is_ergodic(x)

Parameters

x: some kind of Markov chain.

Returns

true if the Markov chain, x, is ergodic. This is, if every state can be accessed from every other state. Another term for this is irreducible.

DiscreteMarkovChains.is_absorbing — Function

is_absorbing(x)

Parameters

x: some kind of Markov chain.

Returns

true if the Markov chain, x, is an absorbing chain. So the process is guarenteed to be absorbed eventually.

Communicaton Functions

DiscreteMarkovChains.communication_classes — Function

communication_classes(x)

Parameters

x: some kind of Markov chain.

Returns

A tuple containing 2 arrays.

This first array contains C arrays which store the states that communicate.

The second array is an array of Bool where the ith value is true if the ith communication class is recurrent.

DiscreteMarkovChains.periodicities — Function

periodicities(x::AbstractDiscreteMarkovChain)

A more advanced version of communication_classes designed for discrete Markov chains. It is the same as communication_classes but it returns periodicities as well.

Parameters

x: some kind of discrete Markov chain.

Returns

A tuple containing 3 arrays.

This first array contains C arrays which store the states that communicate.

The second array is an array of Bool where the ith value is true if the ith communication class is recurrent.

The third array is the periodicity of each communication class.

DiscreteMarkovChains.decompose — Function

decompose(x)

Parameters

x: some kind of Markov chain.

Returns

A tuple of four values

The first value is an array of the new states.

The second value is a matrix of recurrent states to recurrent states.

The third value is a matrix of transient states to recurrent states.

The fourth value is a matrix of transient to transient states.

DiscreteMarkovChains.canonical_form — Function

canonical_form(x)

Reorders the states of the transition matrix of x so that recurrent states appear first and transient states appear last.

Parameters

x: some kind of Markov chain.

Returns

A tuple with the new states and the new transition matrix.

Probability Functions

DiscreteMarkovChains.stationary_distribution — Function

stationary_distribution(x)

Parameters

x: some kind of Markov chain.

Returns

A column vector, w, that satisfies the equation $w'T = w$.

DiscreteMarkovChains.fundamental_matrix — Function

fundamental_matrix(x)

The fundamental matrix of a markov chain is defined to be $(C-I)^{-1}$ where $C$ is the sub-transition matrix that takes transient states to transient states.

Parameters

x: some kind of Markov chain.

Returns

The fundamental matrix of the Markov chain

DiscreteMarkovChains.expected_time_to_absorption — Function

expected_time_to_absorption(x)

Parameters

x: some kind of Markov chain.

Returns

An array where element $i$ is the total number of visits to transient state $i$ before leaving the transient super class.

Note

It is advised to have x in canonical form already. This is to avoid confusion of what states make up each element of the array ouput.

DiscreteMarkovChains.exit_probabilities — Function

exit_probabilities(x)

Parameters

x: some kind of Markov chain.

Returns

An array where element $i, j$ is the probability that transient state $i$ will enter recurrent state $j$ on its first step out of the transient states. That is, $e_{i,j}$.

DiscreteMarkovChains.first_passage_probabilities — Function

first_passage_probabilities(x, t, i=missing, j=missing)

This is the probability that the process enters state $j$ for the first time at time $t$ given that the process started in state $i$ at time 0. That is, $f^{(t)}_{i,j}$. If no i or j is given, then it will return a matrix instead with entries $f^{(t)}_{i,j}$ for i and j in the state space of x.

Parameters

x: some kind of Markov chain.

t: the time to calculate the first passage probability.

i: the state that the prcess starts in.

j: the state that the process must reach for the first time.

Returns

A scalar value or a matrix depending on whether i and j are given.