Probability Functions

Contents

Probability Functions

Contents

Index

Documentation

Index

DiscreteMarkovChains.exit_probabilities

DiscreteMarkovChains.first_passage_probabilities

DiscreteMarkovChains.fundamental_matrix

DiscreteMarkovChains.stationary_distribution

Documentation

DiscreteMarkovChains.stationary_distribution — Function

stationary_distribution(x)

Definitions

A stationary distribution of a Markov chain is a probability distribution that remains unchanged in the Markov chain as time progresses. It is a row vector, $w$ such that its elements sum to 1 and it satisfies $w'T = w'$. $T$ is the one-step transiton matrix of the Markov chain.

In other words, $w$ is invariant by the matrix $T$.

Arguments

x: some kind of Markov chain.

Returns

A column vector, $w$, that satisfies the equation $w'T = w'$.

References

Brilliant.org

DiscreteMarkovChains.fundamental_matrix — Function

fundamental_matrix(x)

Definitions

The fundamental matrix of a markov chain is defined to be $(I-C)^{-1}$ where $C$ is the sub-transition matrix that takes transient states to transient states.

The $(i, j)$th entry of the fundamental matrix is the expected number of times the chain is in state $j$ over the whole process given that the chain started in state $i$.

Arguments

x: some kind of Markov chain.

Returns

The fundamental matrix of the Markov chain

DiscreteMarkovChains.exit_probabilities — Function

exit_probabilities(x)

Arguments

x: some kind of Markov chain.

Returns

An array where element $(i, j)$ is the probability that transient state $i$ will enter recurrent state $j$ on its first step out of the transient states. That is, $e_{i,j}$.

DiscreteMarkovChains.first_passage_probabilities — Function

first_passage_probabilities(x, t, i=missing, j=missing)

Definitions

This is the probability that the process enters state $j$ for the first time at time $t$ given that the process started in state $i$ at time 0. That is, $f^{(t)}_{i,j}$. If no i or j is given, then it will return a matrix instead with entries $f^{(t)}_{i,j}$ for i and j in the state space of x.

Why Do We Use A Slow Algorithm?

So that t can be symbolic if nessesary. That is, if symbolic math libraries want to use this library, it will pose no hassle.

Arguments

x: some kind of Markov chain.

t: the time to calculate the first passage probability.

i: the state that the prcess starts in.

j: the state that the process must reach for the first time.

Returns

A scalar value or a matrix depending on whether i and j are given.

References

University of Windsor

Durham University